requirements.txt
ffmpeg version 7.1.1-1+b1
yt-dlp-2025.4.30